splash package renamed to login
