+++ app/src/main/java/com/ably/tracking/demo/publisher/KoinModule.kt
@@ -26,7 +26,7 @@ import com.ably.tracking.demo.publisher.ui.ActivityNavigator
 import com.ably.tracking.demo.publisher.ui.Navigator
 import com.ably.tracking.demo.publisher.ui.main.MainViewModel
 import com.ably.tracking.demo.publisher.ui.settings.SettingsActionsProvider
-import com.ably.tracking.demo.publisher.ui.splash.LoginViewModel
+import com.ably.tracking.demo.publisher.ui.login.LoginViewModel
 import com.google.gson.GsonBuilder
 import java.util.Locale
 import java.util.TimeZone
