+++ app/src/main/java/com/ably/tracking/demo/publisher/ui/login/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.ably.tracking.demo.publisher.ui.splash
+package com.ably.tracking.demo.publisher.ui.login
 
 import com.ably.tracking.demo.publisher.common.BaseViewModel
 import com.ably.tracking.demo.publisher.secrets.SecretsManager
